First hackish NixOS-ARM-support
Generalized, this does not work.

A newly created NixOS image seems to first have to go through a first boot
before it is appropriately populated. Consequently, modifications to this
image are very difficult. Furthermore, there doesn't seem to be a symlinked
PATH yet, so it has to be created by yourself.

This could be a working Pifile. However, the PATH has not been adjusted
and all in all it is unusable.

> FROM sd-image-armv7l-linux.img
> TO nix.img
>
> RUN /nix/store/w8mmqnbx2ybbax6vkbzm2866m03x8nn4-bash-4.4-p23/bin/bash