PUMP: fix behavior for bigger partition tables
Previously, fdisk was used, which has shown a different UI dialog
depending on how many partitions are already existing. If there was a
possibility for a primary partition, the user was asked. However, if
there are already too much partitions, this dialog was obviously
skipped. This is a pain to automate.

By switching from fdisk to parted, we are now able to resize the
partition regardless of how many are already existing.

To check this, the CI was extended with another check.

This issue was discussed in #34.